Library update + Build parameters
Updated the HumboldtJSLibrary with a number of improvements to the
Animator and EasyStyler class related to opacity. You can now use
"autoOpacity" to have automatic "display:none" added when it reaches 0.
Also added aliases for opacity, autoOpacity, alpha and autoAlpha. This
is implemented in EasyStyler which is now used by Animator, so works
exactly the same.

Also updated the build system to include PureMVC by default.