Updated HumboldtJSLibrary
The update contains a simple HtmlParser utility that can retrieve
elements from a string representation of an HTML file in a DOM-like
fashion

Also added is an XHRLoader which can load either text or XML through an
XMLHttpRequest. It also provides a cross browser XMLHttpRequest factory
method, which deals with all the various browser inconsistencies.

Finally a small update to the URLRequest which allows you to indicate
TEXT or XML as a desired response content type.